pyproject.toml
[project]
name = "enigma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.1.2",
    "plotext>=5.3.2",
    "pyyaml>=6.0.2",
    "rich>=13.9.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]
